ci: add tarpaulin to the devcontainer (#55)
* ci: add tarpaulin to the devcontainer

Changes in preperation for enabling code coverage.

- Create scripts for post create and post attach
for the devcontainer.
- Install tarpaulin in the devcontainer.
- Remove the json pre-commit check as
devcontainer.json is not a properly formmatted
json file.

Part of #12

* ci: install codecov-cli in the devcontainer

Install codecov-cli in the devcontainer so we can
manually upload coverage reports to codecov.io
when unit testing inside the devcontainer.

Had to change the base image to bullseye to work
around an issue installed python modules in
bookworm.

* fix: fix code review comments

Fix cut and paste error in postAttachCommand.sh.