ci: move setup to Dockerfile to speed up builds
The test cycle in the CI pipeline is slow because
the setup is done in the postCreateCommand.sh
script. The work done in the script cannot be
cached by Docker, so it is repeated in every CI
run.

Moving this code to the Dockerfile should allow it
to be cached by Docker, speeding up the CI
pipeline.

Part of #63