ci: remove codecov from devcontainer
In guthub we no longer use the devcontainer for
testing, so we can remove the dependancy on
the codecov cli tool.

With this removed, we can now use a rust
devcontainer rather than a python one. Change
the Dockerfile to use the rust devcontainer and
remove rust from the features section of the
devcontainer.json file.

TEST=devcontainer stats && cargo test

Fixes: #118