fix: do not apply a depth limit when calculating mean in calibrate
When calibrating the mean depth for a region in the `calibrate`
subcommand, we were using a default maximum depth of 8,000 (which is
what htslib uses). If the coverage exceeds this, the downsampling is
incorrect. This commit changes the behaviour to always use the maximum
max depth possible (u32::MAX). This ensures the user will always get a
correctly calibrated BAM file.