feat: adjust calibration algorithm for standard coverage
When calibrating to a standard coverage, the mean coverage was
consistently higher than expected. This was due to an error in the way
reads were selected for keeping or excluding resulting is some
singletons being mistakenly retained. This commit fixes this so that
both pairs of a fragment are always retained, and the resulting mean
coverage is closer to the requested coverage.