feat: calibrate by region mean depth with sample bed (#192)
The calibration command has been updated so that when a sample BED file
is supplied it is used to extract the corresponding mean coverage for
the each sequin, and the coverage is calibrated to that over the entire
sequin region. This is different from the previous algorithm that
attempted to mimic the samples coverage profile. The old behaviour is
still accessible via an `--experimental` argument, but should not be
used in production settings.

When running calibrate, if the argument `--summary-report` is provided,
create a summary CSV with details of the coverage in sequin regions
before and after calibration.