ci: same rust version in devcontainer and workflows
This updates the devcontainer to use a more precisely pinned version of
the container, which also pins the Rust version, and updates the action
that runs outside of the devcontainer to use the same Rust version.

`actions-rs/toolchain` is replaced by `dtolnay/rust-toolchain` as the
former is archived and unmaintained, while the latter is under active
maintanance.