feat: add samtools to docker container (#227)
This commit adds `samtools` to the docker container. It's useful to
have a copy of `samtools` in the `sequintools` container as it can be
used to manipulate the output BAM file in many ways. It also adds a
`.dockerignore` to exclude the target build and test data directories
from the build.

Fixes the docker file which was calling `apt-get clean autoclean`.
While this doesn't cause an error, the `autoclean` doesn't do
anything and is ignored.