refactor: create a library that the binary depends on
This commit creates a library crate along side the main binary crate.
Within the library, proper error types are used and the use of `anyhow`
is eliminated. The `region` module has been placed inside the new
library.

This is the first stage of a multi-stage refactoring with the aim to
improve the code structure to improve our ability to test code and allow
the introduction of a much faster coverage calculation algorithm.