feat: bedcov command uses new coverage module
The bedcov command has been updated to use the library coverage code.
This means that the whole `bedcov` module at the binary level has been
removed. Much of the implementation of `DepthResult` has also been
removed as it was only used by bedcov.

The intent is to also update the `calibrate` command to use the same
coverage calculation code, but it requires more extensive refactoring
before this can happen.