chore: update testdata files
This commit replaces the files under testdata with a set based on an
artifical genome with a full sequins resource bundle. The intent is to
use these files to also provide E2E testing of the binaries in addition
to the unit tests. This has cause the test suite to slow down, but
changing the calibrate code to use the `coverage` module should improve
things.