test: update MockBamReader and add BawWriter
The implementation of `fetch` in `MockBamReader` has been updated to
provide a working version (previously it was just a no-op).

A `BamWriter` trait has been added along with a `HtslibBamWriter` and
`MockBamWriter` implementations. This allows us to mock the writer in
other tests in the crate.

All tests have been updated where required, and tests for the writer
implementation have been added.