feat: calibrate command uses calibration module (#246)
The `calibrate` command now uses the library functions in the
`calibration` module. A new `run_calibrate` function has been added to
`main.rs` to set up the reader/writer, and then calls to the library for
processing. This means that the `calibrate.rs` file is no longer needed
and has been removed.