feat: update cli version message
Include `git describe` in version message produced with `sequintools
--version`. This will now include how many commits from the last tag and
the commit hash used to build the binary. This can be vital in
reproducing bugs or understanding inconsistencies in output. If built
from latest commit, it may be many commits since the last tag, so just
using this is not helpful.