ci: remove .git from .dockerignore
A previous commit updated the way the version is detected in a way that
requires the git directory (`.git`). This was not being sent to the
docker daemon when building containers causing a failure. This commit
removes the git directory from the `.dockerignore` file and makes the
build function more robust when `.git` is missing.