fix: add build.rs to docker build (#253)
Updates the dockerfile and GitHub Action to provide the git derived
version string. A previous commit added this logic to the `build.rs`
file, but it and the `.git` directory are not added to the docker build 
causing it to fail. It's not recommended to add `.git` to a docker
build, so this adds an ARG to the docker files and sets it it the action.