feat: keep all read pairs in calibrate
When calibrating data, it is important to keep ALL reads associated with
a qname that is selected to be kept. There were some edge cases - mainly
with reads mapped to a different chromosome or secondary/supplementary
reads - where this wasn't happening. Downstream tools would fail because
of this.

This commit changes the calibration algorithm so that it maintains a
list of qnames to keep. It then iterates the entire input file to ensure
that all reads associated with the qname are kept. This requires parsing
the file twice: the first time parses just the Sequins decoy chromosome
to determine reads to keep and the second parses the entire file.

See BIOIN-101