fix: don't skip duplicate reads in subsample
Removes the check for duplicate reads in the subsample function. When
calculating sample and sequin coverage, duplicates are not skipped;
therefore, if we skip them when calibrating, we end up with a mean
coverage that is much lower than intended. The subsample function now
considers duplicate reads just like any other.