feat: keep all read pairs in calibrate (#272)
When calibrating data, it is important to keep ALL reads associated with
a qname that is selected to be kept. There were some edge cases - mainly
with reads mapped to a different chromosome or secondary/supplementary
reads - where this wasn't happening. Downstream tools would fail because
of this.

This commit changes the calibration algorithm so that it maintains a
list of qnames to keep. It then iterates the entire input file to ensure
that all reads associated with the qname are kept. This requires parsing
the file twice: the first time parses just the Sequins decoy chromosome
to determine reads to keep and the second parses the entire file.

Removes the check for duplicate reads in the subsample function. When
calculating sample and sequin coverage, duplicates are not skipped;
therefore, if we skip them when calibrating, we end up with a mean
coverage that is much lower than intended. The subsample function now
considers duplicate reads just like any other.

See BIOIN-101