feat: add PG record to alignment output files
When creating BAM/CRAM files add a PG (Program) record. This record
records the exact command line and version of sequintools used to
produce the output. This can be invaluable when analysing data (or
bugs).

As part of this, the way the integration tests check the MD5 of the
output was changed. Instead computing the MD5 on the output file, it is
first filtered to remove the PG records. This is because they record the
exact command line - with the temporary directory which changes every
time the tests are run. This causes the checksum to change every time
too. By removing the PG lines, we get deterministic output.

Resolves BIOIN-114